+++ plugins/main/common/constants.ts
@@ -564,6 +564,7 @@ export enum WAZUH_MODULES_ID {
   PCI_DSS = 'pci',
   HIPAA = 'hipaa',
   NIST_800_53 = 'nist',
+  NIST_800_171 = 'nist-800-171',
   TSC = 'tsc',
   NIS2 = 'nis2',
   VIRUSTOTAL = 'virustotal',
@@ -611,6 +612,8 @@ export const DATA_SOURCE_FILTER_CONTROLLED_TSC_EXIST = 'tsc-rule-exist';
 export const DATA_SOURCE_FILTER_CONTROLLED_CMMC_EXIST = 'cmmc-rule-exist';
 export const DATA_SOURCE_FILTER_CONTROLLED_NIST_800_53_EXIST =
   'nist-800-53-rule-exist';
+export const DATA_SOURCE_FILTER_CONTROLLED_NIST_800_171_EXIST =
+  'nist-800-171-rule-exist';
 export const DATA_SOURCE_FILTER_CONTROLLED_GDPR_EXIST = 'gdpr-rule-exist';
 export const DATA_SOURCE_FILTER_CONTROLLED_HIPAA_EXIST = 'hipaa-rule-exist';
 export const DATA_SOURCE_FILTER_CONTROLLED_NIS2_EXIST = 'nis2-rule-exist';
@@ -1021,6 +1024,10 @@ export const HIPAA_AGENT_DASHBOARD_ID = 'hipaa-pinned-agent-dashboard-tab';
 
 export const NIST_800_53_DASHBOARD_ID = 'nist-overview-dashboard-tab';
 export const NIST_800_53_AGENT_DASHBOARD_ID = 'nist-pinned-agent-dashboard-tab';
+
+export const NIST_800_171_DASHBOARD_ID = 'nist-800-171-overview-dashboard-tab';
+export const NIST_800_171_AGENT_DASHBOARD_ID =
+  'nist-800-171-pinned-agent-dashboard-tab';
 
 export const TSC_DASHBOARD_ID = 'tsc-overview-dashboard-tab';
 export const TSC_AGENT_DASHBOARD_ID = 'tsc-pinned-agent-dashboard-tab';