+++ plugins/main/public/components/overview/regulatory-compliance/nist-800-171/events/nist-171-columns.tsx
@@ -1,0 +1,14 @@
+import { tDataGridColumn } from '../../../../common/data-grid';
+import { commonColumns } from '../../../common/data-grid-columns';
+
+export const nist171Columns: tDataGridColumn[] = [
+  commonColumns.timestamp,
+  commonColumns['wazuh.agent.name'],
+  {
+    id: 'rule.compliance.nist_800_171',
+    initialWidth: 171,
+  },
+  { id: 'wazuh.integration.name' },
+  { id: 'wazuh.integration.category' },
+  { id: 'wazuh.integration.decoders' },
+];